Update README.md
Like most meta-heuristic algorithms, Salp Swarm Optimization (SSA) suffers from slow convergence and stagnation in the local optima . The study develops a novel Time-Based Leadership Salp-Based Competitive Learning (TBLSBCL) to address the SSA's flaws. The TBLSBCL presents a novel search technique to address population diversity, an imbalance between exploitation and exploration, and the SSA algorithm's premature convergence. Hybridization consists of two stages: First, a time-varying dynamic structure represents the SSA hierarchy of leaders and followers. This approach increases the number of leaders while decreasing the number of salp’s followers linearly. Utilizing the effective exploitation of the SSA, the position of the population's leader is updated. Second, the Competitive Learning Strategy is utilized to update the status of the followers by teaching them from the leaders. The goal of adjusting the Salp swarm optimizer algorithm is to help the basic approach avoid premature convergence and quickly steer the search to the most promising likely search space. The proposed TBLSBCL method is tested using the CEC 2017 benchmark, Feature Selection problems for 19 datasets (including three high-dimensional datasets). The TBLSBCL was then evaluated using a benchmark set of seven well-known constrained design challenges in diverse engineering fields defined in the benchmark set of real-world problems presented at the CEC 2020 conference (CEC 2020).  In each experiment, TBLSBCL is compared with seven other state-of-the-art metaheuristics and other advanced algorithms that include seven variants of the Salp swarm. Friedman and Wilcoxon rank-sum statistical tests are also used to examine the results. According to the experimental data and statistical tests, the TBLSBCL algorithm is very competitive and often superior to the algorithms employed in the studies